feat(frontend): move beforeunload functionality to hook bb-473
