+++ apps/frontend/src/pages/settings/settings.tsx
@@ -15,7 +15,6 @@ import {
 	useAppForm,
 	useAppSelector,
 	useCallback,
-	useEffect,
 	useUnsavedChangesBlocker,
 } from "~/libs/hooks/hooks.js";
 import {
@@ -65,24 +64,6 @@ const Settings: React.FC = () => {
 
 	const { blockerState, handlePopupCancel, handlePopupConfirm } =
 		useUnsavedChangesBlocker({ hasUncavedChanges: isDirty, reset });
-
-	const handleBeforeUnload = useCallback(
-		(event: BeforeUnloadEvent): void => {
-			if (isDirty) {
-				event.preventDefault();
-				event.returnValue = "Reload site?";
-			}
-		},
-		[isDirty],
-	);
-
-	useEffect(() => {
-		window.addEventListener("beforeunload", handleBeforeUnload);
-
-		return (): void => {
-			window.removeEventListener("beforeunload", handleBeforeUnload);
-		};
-	}, [handleBeforeUnload]);
 
 	return (
 		<>
